Refactor attendance, billing, and facility modules; remove unused classes and restructure domain entities
