+++ src/main/java/com/careconnect/coreapi/attendance/domain/package-info.java
@@ -1,0 +1,7 @@
+/**
+ * Domain entities for the attendance module.
+ * 
+ * This package contains the core domain entities that represent attendance tracking concepts.
+ * These entities should not be directly accessed by other modules.
+ */
+package com.careconnect.coreapi.attendance.domain;
