+++ src/main/java/com/careconnect/coreapi/billing/domain/package-info.java
@@ -1,0 +1,7 @@
+/**
+ * Domain entities for the billing module.
+ * 
+ * This package contains the core domain entities that represent billing and payment concepts.
+ * These entities should not be directly accessed by other modules.
+ */
+package com.careconnect.coreapi.billing.domain;
