--- src/main/java/com/careconnect/coreapi/facility/FacilityInfo.java
@@ -1,16 +1,0 @@
-package com.careconnect.coreapi.facility;
-
-import java.util.UUID;
-
-/**
- * Public API for Facility information exposed to other modules
- */
-public record FacilityInfo(
-    UUID id,
-    String name,
-    String phone,
-    String licenseNumber,
-    Integer maxCapacity
-) {
-    
-}
