+++ src/main/java/com/careconnect/coreapi/facility/domain/package-info.java
@@ -1,0 +1,7 @@
+/**
+ * Domain entities for the facility module.
+ * 
+ * This package contains the core domain entities that represent facility and address concepts.
+ * These entities should not be directly accessed by other modules.
+ */
+package com.careconnect.coreapi.facility.domain;
