+++ src/main/java/com/careconnect/coreapi/reports/domain/package-info.java
@@ -1,0 +1,7 @@
+/**
+ * Domain entities for the reports module.
+ * 
+ * This package contains the core domain entities that represent reporting concepts.
+ * These entities should not be directly accessed by other modules.
+ */
+package com.careconnect.coreapi.reports.domain;
